+++ physical/physical.go
@@ -128,7 +128,7 @@ var builtinBackends = map[string]Factory{
 	"mysql":      newMySQLBackend,
 	"postgresql": newPostgreSQLBackend,
 	"swift":      newSwiftBackend,
-	"gcs":     newGCSBackend,
+	"gcs":        newGCSBackend,
 }
 
 // PermitPool is used to limit maximum outstanding requests
